Normalization Methods for MCDM

In MCDM methods, e.g. topsis, aras, waspas, each single method has a predefined and default normalization method. However the normalization method can be specifically defined by setting the normalization= parameter in MCDM functions.

For example the default Topsis call with Vector norm normalization is

topsis(data, weights, fns)

whereas, the same method can be called with a different normalization method like

topsis(data, weights, fns, normalization = Normalizations.dividebycolumnsumnormalization

Vector Norm Normalization

JMcDM.Normalizations.vectornormnormalization — Function

vectornormnormalization(data::Matrix, fns)::Matrix

Default normalization method for Topsis and Moora

Divide by Column Sums Normalization

JMcDM.Normalizations.dividebycolumnsumnormalization — Function

dividebycolumnsumnormalization(data::Matrix, fns)::Matrix

Default normalization method for Aras and Copras

Max-Min Range Normalization

JMcDM.Normalizations.maxminrangenormalization — Function

maxminrangenormalization(data::Matrix, fns)::Matrix

Default normalization method for Cocoso, Critic, Grey, Mabac, and Mairca

Inverse Max-Min Range Normalization (max->min, min->max)

JMcDM.Normalizations.inversemaxminrangenormalization — Function

inversemaxminrangenormalization(data::Matrix, fns)::Matrix

Grouped Max-Min Range Normalization

JMcDM.Normalizations.groupeddividebymaxminnormalization — Function

groupeddividebymaxminnormalization(mat::Matrix, fns)

Default normalization method for seca.

Divide by Column Maximum-Minimum Normalization

JMcDM.Normalizations.dividebycolumnmaxminnormalization — Function

dividebycolumnmaxminnormalization(mat::Matrix, fns)

Default normalization method for Codas and Psi.

Inverse Divide by Column Maximum-Minimum Normalization

JMcDM.Normalizations.inversedividebycolumnmaxminnormalization — Function

inversedividebycolumnmaxminnormalization(mat::Matrix, fns)

Default normalization method by Merec.

Divide by All Norm Normalization

JMcDM.Normalizations.dividebyallnormnormalization — Function

dividebyallnormnormalization(mat::Matrix, fns)

Default normalization method for Moosra

Null Normalization

JMcDM.Normalizations.nullnormalization — Function

nullnormalization(data::Matrix, fns)::Matrix

Default normalization method for disabling normalization