Single-criterion decision making tools

laplace

JMcDM.SCDM.laplace — Function

laplace(decisionMat)

Apply Laplace method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

Output

::LaplaceResult: LaplaceResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        3000 2750 2500 2250;
        1500 4750 8000 7750;
        2000 5250 8500 11750
]

julia> result = laplace(mat)

maximin

JMcDM.SCDM.maximin — Function

maximin(decisionMat)

Apply Maximin method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

Output

::MaximinResult: MaximinResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> result = maximin(mat)

maximax

JMcDM.SCDM.maximax — Function

maximax(decisionMat)

Apply Maximax method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

Output

::MaximaxResult: MaximaxResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> result = maximax(mat)

minimax

JMcDM.SCDM.minimax — Function

minimax(decisionMat)

Apply Minimax method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

Output

::MinimaxResult: MinimaxResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> result = minimax(mat)

minimin

JMcDM.SCDM.minimin — Function

minimin(decisionMat)

Apply Minimin method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

Output

::MiniminResult: Minimin object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> result = minimin(mat)

savage

JMcDM.SCDM.savage — Function

savage(decisionMat)

Apply Savage method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

Output

::SavageResult: SavageResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> result = savage(mat)

julia> result.bestIndex 
4

hurwicz

JMcDM.SCDM.hurwicz — Function

hurwicz(decisionMat; alpha = 0.5)

Apply Hurwicz method for a given decision matrix (for convenience, in type of Matrix).

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

alpha::Float64: The optional alpha value for the Hurwicz method. Default is 0.5.

Output

::HurwiczResult: HurwiczResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> result = hurwicz(mat)

julia> result.bestIndex 
3

mle

JMcDM.SCDM.mle — Function

mle(decisionMat, weights)

Apply MLE (Maximum Likelihood) method for a given decision matrix (for convenience, in type of Matrix) and weights.

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

weights::Array{Float64,1}: Array of weights for each criterion that sums up to 1.0.

Output

::MLEResult: MLEResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> weights = [0.2, 0.5, 0.2, 0.1]
julia> result = mle(mat, weights)

julia> result.bestIndex 
2

expectedregret

JMcDM.SCDM.expectedregret — Function

expectedregret(decisionMat, weights)

Apply Expected Regret method for a given decision matrix (for convenience, in type of Matrix) and weights.

Arguments:

decisionMat::Matrix: Decision matrix with n alternatives and m criteria.

weights::Array{Float64,1}: Array of weights for each criterion that sums up to 1.0.

Output

::ExpectedRegretResult: ExpectedRegretResult object that holds multiple outputs including the best alternative.

Examples

julia> mat = [
        26 26 18 22;
        22 34 30 18;
        28 24 34 26;
        22 30 28 20
    ]

julia> weights = [0.2, 0.5, 0.2, 0.1]
julia> result = expectedregret(mat, weights)

julia> result.bestIndex 
2